MODIFY inputs inside forms

+++ admin/liste-employes.php
@@ -8,12 +8,6 @@ require_once __DIR__ . "/../lib/pdo.php";
 require_once __DIR__ . "/../lib/employees.php";
 require_once __DIR__ . "/templates/header-admin.php";
 
-// //only admin can visit this page
-// if ($_SESSION['user']['role'] === 'employee') {
-//   header("location: /admin/liste-voitures.php");
-// }
-
-
 if (isset($_GET['page'])) {
   $page = (int)$_GET['page'];
 } else {
@@ -51,21 +45,19 @@ $totalPages = ceil($totalEmployees / _ADMIN_ITEM_PER_PAGE_);
           </tr>
           <tbody>
             <?php foreach ($employees as $employee) { ?>
-            <tr>
-              <!-- <th scope="row"><?= $employee["id"]; ?></th> -->
-              <td><?= $employee["lastname"]; ?></td>
-              <td><?= $employee["name"]; ?></td>
-              <td><?= $employee["email"]; ?></td>
-              <td><?= $employee["role"]; ?></td>
-              <?php if($employee["role"] === 'employee'){?>
+              <tr>
+                <!-- <th scope="row"><?= $employee["id"]; ?></th> -->
+                <td><?= $employee["lastname"]; ?></td>
+                <td><?= $employee["name"]; ?></td>
+                <td><?= $employee["email"]; ?></td>
+                <td><?= $employee["role"]; ?></td>
 
-              <td><a href="modifier-employe.php?id=<?= $employee['id'] ?>"><i class="fa-solid fa-pencil"></i></a>
-                | <a href="delete-employe.php?id=<?= $employee['id'] ?>"
-                  onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet employé ?')"><i
-                    class="fa-solid fa-trash-can"></i></a>
-              </td>
-              <?php } ?>
-            </tr>
+                <td><a href="modifier-employe.php?id=<?= $employee['id'] ?>"><i class="fa-solid fa-pencil"></i></a>
+                  <?php if ($employee["role"] === 'employee') { ?>
+                    | <a href="delete-employe.php?id=<?= $employee['id'] ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet employé ?')"><i class="fa-solid fa-trash-can"></i></a>
+                  <?php } ?>
+                </td>
+              </tr>
             <?php } ?>
           </tbody>
         </table>
@@ -74,13 +66,15 @@ $totalPages = ceil($totalEmployees / _ADMIN_ITEM_PER_PAGE_);
       <nav aria-label="Page navigation employes">
         <ul class="pagination">
           <?php if ($totalPages > 1) { ?>
-          <?php for ($i = 1; $i <= $totalPages; $i++) { ?>
-          <li class="page-item">
-            <a class="page-link <?php if ($i == $page) { echo " active";} ?>" href="?page=<?php echo $i; ?>">
-              <?php echo $i; ?>
-            </a>
-          </li>
-          <?php } ?>
+            <?php for ($i = 1; $i <= $totalPages; $i++) { ?>
+              <li class="page-item">
+                <a class="page-link <?php if ($i == $page) {
+                                      echo " active";
+                                    } ?>" href="?page=<?php echo $i; ?>">
+                  <?php echo $i; ?>
+                </a>
+              </li>
+            <?php } ?>
           <?php } ?>
         </ul>
       </nav>
